+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to the "cosmy-wasmy" extension will be documented in this fi
 - Support for CW notebooks - Can instantiate contract, run queries and execute txs from the notebook cell within the cosmwasm virtual machine kernel.
 - Added support for `SigningCosmWasmClient.executeMultiple` in API
 - Can now send funds while executing a contract and while initializing a contract 💰
+- Added `Clear History` button and `Export History as JSON` button to CosmWasm history view
 
 ### Changed